# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Warehouse AI Video Intelligence API",
    "description": "Warehouse handling video intelligence API prototype",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/runs": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Create Run",
        "description": "Atomically enqueue analysis run and background job.",
        "operationId": "create_run_api_v1_runs_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunCreate"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Runs",
        "description": "List runs ordered by creation date descending.",
        "operationId": "get_runs_api_v1_runs_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "video_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Video Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RunDetail"
                  },
                  "title": "Response Get Runs Api V1 Runs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Run By Id",
        "description": "Retrieve run details by ID.",
        "operationId": "get_run_by_id_api_v1_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/events": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Get Events",
        "description": "List events matching filters with cursor pagination.",
        "operationId": "get_events_api_v1_events_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          },
          {
            "name": "event_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Event Type"
            }
          },
          {
            "name": "risk_tier",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Risk Tier"
            }
          },
          {
            "name": "review_verdict",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Review Verdict"
            }
          },
          {
            "name": "start_ms",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Ms"
            }
          },
          {
            "name": "end_ms",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Ms"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/events/{event_id}": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Get Event By Id",
        "description": "Retrieve event details, facts, trace, and media links by ID.",
        "operationId": "get_event_by_id_api_v1_events__event_id__get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/events/{event_id}/reviews": {
      "post": {
        "tags": [
          "events"
        ],
        "summary": "Post Event Review",
        "description": "Append a human review revision for an event.",
        "operationId": "post_event_review_api_v1_events__event_id__reviews_post",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EventDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "event_type": {
            "type": "string",
            "enum": [
              "DROP",
              "FORCEFUL_RELEASE",
              "DRAGGING",
              "VISIBLE_SUPPORT_OVERHANG",
              "PROHIBITED_ZONE",
              "STANDING_ON_PRODUCT"
            ],
            "title": "Event Type"
          },
          "start_ms": {
            "type": "integer",
            "title": "Start Ms"
          },
          "end_ms": {
            "type": "integer",
            "title": "End Ms"
          },
          "primary_track_id": {
            "type": "integer",
            "title": "Primary Track Id"
          },
          "risk": {
            "$ref": "#/components/schemas/RiskDetail"
          },
          "evidence_quality": {
            "type": "number",
            "title": "Evidence Quality"
          },
          "verification_status": {
            "type": "string",
            "enum": [
              "NOT_RUN",
              "PASSED",
              "FAILED"
            ],
            "title": "Verification Status"
          },
          "facts": {
            "additionalProperties": true,
            "type": "object",
            "title": "Facts"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          },
          "media": {
            "$ref": "#/components/schemas/MediaLinks"
          },
          "review": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReviewResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "run_id",
          "event_type",
          "start_ms",
          "end_ms",
          "primary_track_id",
          "risk",
          "evidence_quality",
          "verification_status",
          "facts",
          "explanation",
          "media"
        ],
        "title": "EventDetail"
      },
      "EventPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EventDetail"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "EventPage"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MediaLinks": {
        "properties": {
          "clip_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clip Id"
          },
          "thumbnail_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "MediaLinks"
      },
      "ReviewCreate": {
        "properties": {
          "verdict": {
            "type": "string",
            "enum": [
              "CONFIRMED_RISK",
              "REJECTED",
              "UNCERTAIN"
            ],
            "title": "Verdict"
          },
          "note": {
            "type": "string",
            "maxLength": 1000,
            "title": "Note",
            "default": ""
          },
          "reviewer_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Reviewer Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "verdict",
          "reviewer_name"
        ],
        "title": "ReviewCreate"
      },
      "ReviewResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "verdict": {
            "type": "string",
            "enum": [
              "CONFIRMED_RISK",
              "REJECTED",
              "UNCERTAIN"
            ],
            "title": "Verdict"
          },
          "note": {
            "type": "string",
            "title": "Note"
          },
          "reviewer_name": {
            "type": "string",
            "title": "Reviewer Name"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "event_id",
          "revision",
          "verdict",
          "note",
          "reviewer_name",
          "created_at"
        ],
        "title": "ReviewResponse"
      },
      "RiskDetail": {
        "properties": {
          "score": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Score"
          },
          "tier": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "CRITICAL"
            ],
            "title": "Tier"
          },
          "policy_version": {
            "type": "string",
            "title": "Policy Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "score",
          "tier",
          "policy_version"
        ],
        "title": "RiskDetail"
      },
      "RunAccepted": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "const": "QUEUED",
            "title": "Status",
            "default": "QUEUED"
          },
          "status_url": {
            "type": "string",
            "title": "Status Url"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "job_id",
          "status_url"
        ],
        "title": "RunAccepted"
      },
      "RunCreate": {
        "properties": {
          "video_id": {
            "type": "string",
            "title": "Video Id"
          },
          "camera_profile_id": {
            "type": "string",
            "title": "Camera Profile Id"
          },
          "camera_profile_version": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Camera Profile Version"
          },
          "mode": {
            "type": "string",
            "enum": [
              "LIVE",
              "REPLAY"
            ],
            "title": "Mode",
            "default": "LIVE"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "video_id",
          "camera_profile_id",
          "camera_profile_version"
        ],
        "title": "RunCreate"
      },
      "RunDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "video_id": {
            "type": "string",
            "title": "Video Id"
          },
          "camera_profile_id": {
            "type": "string",
            "title": "Camera Profile Id"
          },
          "camera_profile_version": {
            "type": "integer",
            "title": "Camera Profile Version"
          },
          "mode": {
            "type": "string",
            "enum": [
              "LIVE",
              "REPLAY"
            ],
            "title": "Mode"
          },
          "status": {
            "type": "string",
            "enum": [
              "QUEUED",
              "RUNNING",
              "SUCCEEDED",
              "FAILED"
            ],
            "title": "Status"
          },
          "model_backend": {
            "type": "string",
            "title": "Model Backend"
          },
          "rules_version": {
            "type": "string",
            "title": "Rules Version"
          },
          "risk_policy_version": {
            "type": "string",
            "title": "Risk Policy Version"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "manifest_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Manifest Url"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "video_id",
          "camera_profile_id",
          "camera_profile_version",
          "mode",
          "status",
          "model_backend",
          "rules_version",
          "risk_policy_version",
          "warnings",
          "created_at"
        ],
        "title": "RunDetail"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}